feat: delete and unify cosmos send form modal
