+++ src/components/Modals/Send/Send.tsx
@@ -1,8 +1,6 @@
 import { Modal, ModalContent, ModalOverlay } from '@chakra-ui/react'
 import type { Asset } from '@shapeshiftoss/asset-service'
 import type { AccountId } from '@shapeshiftoss/caip'
-import { CHAIN_NAMESPACE, fromChainId } from '@shapeshiftoss/caip'
-import { Form as CosmosForm } from 'plugins/cosmos/components/modals/Send/Form'
 import { useRef } from 'react'
 import type { RouteComponentProps } from 'react-router-dom'
 import { MemoryRouter, Route, Switch } from 'react-router-dom'
@@ -37,14 +35,9 @@ export const SendModal = ({ asset, accountId }: SendModalProps) => {
           <Switch>
             <Route
               path='/'
-              component={(props: RouteComponentProps) => {
-                const { chainNamespace } = fromChainId(asset.chainId)
-                return chainNamespace === CHAIN_NAMESPACE.CosmosSdk ? (
-                  <CosmosForm asset={asset} accountId={accountId} {...props} />
-                ) : (
-                  <Form asset={asset} accountId={accountId} {...props} />
-                )
-              }}
+              component={(props: RouteComponentProps) => (
+                <Form asset={asset} accountId={accountId} {...props} />
+              )}
             />
           </Switch>
         </MemoryRouter>
