+++ src/components/Modals/Send/views/Details.tsx
@@ -7,16 +7,21 @@ import {
   FormHelperText,
   FormLabel,
   IconButton,
+  Input,
   ModalBody,
   ModalCloseButton,
   ModalFooter,
   ModalHeader,
   Stack,
+  Tooltip,
 } from '@chakra-ui/react'
 import type { AccountId } from '@shapeshiftoss/caip'
+import { fromAssetId } from '@shapeshiftoss/caip'
+import { CHAIN_NAMESPACE } from '@shapeshiftoss/caip/dist/constants'
 import isNil from 'lodash/isNil'
-import { useCallback } from 'react'
-import { useFormContext, useWatch } from 'react-hook-form'
+import { useCallback, useMemo } from 'react'
+import { Controller, useFormContext, useWatch } from 'react-hook-form'
+import { FaInfoCircle } from 'react-icons/fa'
 import { useTranslate } from 'react-polyglot'
 import { useHistory } from 'react-router-dom'
 import { AccountCard } from 'components/AccountCard'
@@ -27,21 +32,32 @@ import { Text } from 'components/Text'
 import { TokenRow } from 'components/TokenRow/TokenRow'
 import { useModal } from 'hooks/useModal/useModal'
 import { useWallet } from 'hooks/useWallet/useWallet'
+import { bnOrZero } from 'lib/bignumber/bignumber'
 
 import type { SendInput } from '../Form'
 import { useSendDetails } from '../hooks/useSendDetails/useSendDetails'
 import { SendFormFields, SendRoutes } from '../SendCommon'
 import { SendMaxButton } from '../SendMaxButton/SendMaxButton'
 
+const MAX_COSMOS_SDK_MEMO_LENGTH = 256
+
 export const Details = () => {
   const { control, setValue } = useFormContext<SendInput>()
   const history = useHistory()
   const translate = useTranslate()
 
-  const { asset, accountId, cryptoAmount, cryptoSymbol, fiatAmount, fiatSymbol, amountFieldError } =
-    useWatch({
-      control,
-    }) as Partial<SendInput>
+  const {
+    accountId,
+    amountFieldError,
+    asset,
+    cryptoAmount,
+    cryptoSymbol,
+    fiatAmount,
+    fiatSymbol,
+    memo,
+  } = useWatch({
+    control,
+  }) as Partial<SendInput>
 
   const handleAccountChange = useCallback(
     (accountId: AccountId) => {
@@ -68,6 +84,16 @@ export const Details = () => {
   const {
     state: { wallet },
   } = useWallet()
+
+  const showMemoField = useMemo(
+    () => Boolean(asset && fromAssetId(asset.assetId).chainNamespace === CHAIN_NAMESPACE.CosmosSdk),
+    [asset],
+  )
+  const remainingMemoChars = useMemo(
+    () => bnOrZero(MAX_COSMOS_SDK_MEMO_LENGTH - Number(memo?.length)),
+    [memo],
+  )
+  const memoFieldError = remainingMemoChars.lt(0) && 'Characters Limit Exceeded'
 
   if (!(asset && !isNil(cryptoAmount) && cryptoSymbol && !isNil(fiatAmount) && fiatSymbol)) {
     return null
@@ -188,12 +214,67 @@ export const Details = () => {
             />
           )}
         </FormControl>
+        {showMemoField && (
+          <FormControl mt={6}>
+            <Box display='flex' alignItems='center' justifyContent='space-between'>
+              <FormLabel color='gray.500' display='flex' alignItems='center'>
+                <Text
+                  translation={['modals.send.sendForm.assetMemo', { assetSymbol: asset.symbol }]}
+                />
+                <Tooltip
+                  placement='right'
+                  label={translate('modals.send.sendForm.memoExplainer', {
+                    assetSymbol: asset.symbol,
+                  })}
+                  fontSize='md'
+                  pr={4}
+                >
+                  <Box ml='5px'>
+                    <FaInfoCircle />
+                  </Box>
+                </Tooltip>
+              </FormLabel>
+              <FormHelperText
+                mt={0}
+                mr={3}
+                mb={2}
+                as='button'
+                type='button'
+                color={memoFieldError ? 'red.500' : 'gray.500'}
+              >
+                {translate('modals.send.sendForm.charactersRemaining', {
+                  charactersRemaining: remainingMemoChars.toString(),
+                })}
+              </FormHelperText>
+            </Box>
+            <Controller
+              name={SendFormFields.Memo}
+              render={({ field: { onChange, value } }) => (
+                <Input
+                  size='lg'
+                  onChange={({ target: { value } }) => onChange(value)}
+                  value={value}
+                  type='text'
+                  variant='filled'
+                  placeholder={translate('modals.send.sendForm.optionalAssetMemo', {
+                    assetSymbol: asset.symbol,
+                  })}
+                />
+              )}
+            />
+          </FormControl>
+        )}
       </ModalBody>
       <ModalFooter>
         <Stack flex={1}>
           <Button
             width='full'
-            isDisabled={!(cryptoAmount ?? fiatAmount) || !!amountFieldError || loading}
+            isDisabled={
+              !(cryptoAmount ?? fiatAmount) ||
+              !!amountFieldError ||
+              loading ||
+              Boolean(memoFieldError)
+            }
             colorScheme={amountFieldError ? 'red' : 'blue'}
             size='lg'
             onClick={handleNextClick}
